Bug 1675159 - Pt 2 - Expose dirty rects to start_compositing. r=sotaro
Move the calculation of the dirty rects array earlier in frame
drawing, and supply that to the `start_compositing` method of
the compositor trait.

For now, it's assumed that the native compositor wants a single
dirty rect, and doesn't use buffer-age functionality. These
params will be configurable as part of the compositor capabilities
struct in follow up patches.

Differential Revision: https://phabricator.services.mozilla.com/D95828

UltraBlame original commit: 93e746226b8feb5a5e6a643037ef34496caa26bf