Bug 1675437 - Add application name to the crash reporter bundle display name. r=Gijs
Back in bug 382762, it was originally set to contain the application
name, and the Makefile parts have remained, but the application name was
removed in later iterations because back then, the crash reporter would
be shipped with XulRunner and that would be a problem for XulRunner
apps.

XulRunner is long dead, so we can go back to the original idea.

The real motive behind this is that InfoPlist.strings.in needs to be
converted to UTF-16 at build time, and that while we currently do that
with iconv, we want to change that, and a convenient way to do so is to
use the preprocessor, but then the preprocessor doesn't like that this
specific InfoPlist.strings.in doesn't have any preprocessing directives,
so this adds one.

Differential Revision: https://phabricator.services.mozilla.com/D96012

UltraBlame original commit: 415a591e6b75278324d5477a33cae1c74488c1a2