Bug 1674722 - Fix moz.configure setup for cranelift. r=rhunt
We want to enable cranelift on aarch64 hardware except on Windows (due
to ABI issues) and on x64 with the aarch64 simulator, but not on x64
in general (because it is immature).

Add a test case to ensure that cranelift is not enabled where we do
not expect it to be.

Differential Revision: https://phabricator.services.mozilla.com/D96058

UltraBlame original commit: c4511a3c1bc64210f293ffe88ce20a4b0603a5a0