Bug 1669414 - osclientcerts: temporarily remove differentiation between modern and legacy keys r=rmf
Eventually it needs to be possible for osclientcerts to differentiate between
keys that can and can't perform modern cryptography (RSA-PSS being the main
issue). This is because PSM and NSS need to know not to offer to use a key that
can't actually perform the signing operation in question. However, the current
implementation can be very slow if the user has slow hardware with a number of
keys on it. Since PSM and NSS changes are required to make use of this
differentiation anyway, the best approach for now seems to be to skip this step.

Differential Revision: https://phabricator.services.mozilla.com/D96148

UltraBlame original commit: 0ed010f9a8c03231403a18993f5a53d317fcdc8c