Bug 1673634: custom margins must factor in page orientation when calculating maximum values r=mstriemer
When the orientation changes, we must recalculate the maximum height and width.
For portrait orientation, marginTop and marginBottom should be compared to the
paper width, and for landscape orientation, they should be compared to the paper
height (vice versa for marginLeft and marginRight).

If the custom margin values are valid after an orientation switch, they will
persist. If they are no longer valid, they will be reset to the default values,
and the margin dropdown will still have custom selected.

Differential Revision: https://phabricator.services.mozilla.com/D96099

UltraBlame original commit: 1817390a4cbf3241a68a4cf4eef40003702dd093