Bug 1675384 - Resurrect mac unification code. r=firefox-build-system-reviewers,mhentges
This adds back part of the code that was removed in bug 1339182,
reformats it with black, adjusts it to make flake8 happy, and converts
it to python 3.

This also adjusts the script after bug 1534003, which changed the
about:buildconfig page title.

Differential Revision: https://phabricator.services.mozilla.com/D95977

UltraBlame original commit: 9fc9b36bf25945f59c9d2db14668d8c6a05606ae