Bug 1675790 - Issue a soft update correctly for non-subresource requests r=dom-workers-and-storage-reviewers,asuth
The Handle Fetch algorithm, steps 17 and 18, requires a soft update
when a fetch event is not dispatch for a non-subresource request.
Before this commit we added the requirement that the registration
be stale, which only applies to subresource requests.

Differential Revision: https://phabricator.services.mozilla.com/D96200

UltraBlame original commit: ac207bcbee104d1e1f0d11a327d3b9ecb39592f7