Bug 1667778 - [devtools] Remove mentions of launchpad in debugger. r=jdescottes.
This removes the few mentions of the launchpad that were
still existing in the debugger codebase.
We take this as an opportunity to use the #mount element
as the root of the React app (instead of creating a new
div for it).

Depends on D96384

Differential Revision: https://phabricator.services.mozilla.com/D96513

UltraBlame original commit: 217f4a05b021c0c5a2fb2a3d1491634bac573248