Bug 1673391 - Specialize binary SIMD ops with a constant argument.  r=jseward
Introduce a specialization in the front end that rewrites SIMD
binary-op nodes as binary-op-with-constant nodes when one operand is a
constant that is used only the once (either operand for commutative
operations; the rhs for others).  The binary-op-with-constant is then
emitted either as an instruction with an in-memory constant, or with a
synthesized constant in the scratch register, if that is believed to
be faster (using existing logic for that).

Even when the constant can't be synthesized, the rewrite will lower
the register pressure.

This only specializes operations that have straightforward renderings
on the hardware.  See the comment in Lowering-x86-shared.cpp.  More
can, and will, be done.

Most notable by their absence are floating-point
compare-with-constant.  These have straightforward renderings on the
hardware but our assembler has limitations around patching
instructions where the patch location is not at the end of the
instruction, and we need to do that for the float compares with a
rip-relative value.  Thus these ops have been deferred to a later patch.

Differential Revision: https://phabricator.services.mozilla.com/D94806

UltraBlame original commit: 60877a8a79803c764de4c633fe023494a8869db2