Bug 1672072 - P7. Remove the guarantee that the TrackInfo reference will always be valid. r=mattwoodrow,bryce,padenot
Until now, it was up to the caller to ensure that the VideoInfo or AudioInfo passed to the PlatformDecoderModule would remain alive until the created decoder was shut down.
This was achieved by the MediaFormatReader through great care and complexity and ensuring that the MediaFormatReader would never complete its own shutdown unless the  decoder did the same.

Asynchronously creating the decoder add an even greater of complexity. We have to ensure that should the caller shut down while the decoder creation promise hasn't been resolved ; that it stays alive the entire time.

It could be achieved by making the TrackInfo object refcounted, or creating a cycle between the caller and the PDM/decoder.

Each of those options have their own downsides.

Either of those options add an unwarranted level of complexity, when we can just have each decoder copy the arguments it receives. This is simpler and more straightforward and the required copy isn't expensive.

Differential Revision: https://phabricator.services.mozilla.com/D96363

UltraBlame original commit: 09adad7416e19b7fb06f95b26b23818f3b71fc87