Bug 1672072 - P4. Pass lambda rather than raw pointer to object. r=mattwoodrow,bryce,padenot
Passing a pointer to an object across several thread was dubious at best to ensure a use after free was never in place.
This complexity becomes even greater should we want to create the decoder asynchronously.

So we pass instead a lambda that holds a strong reference to the object that holds the event producer.

Differential Revision: https://phabricator.services.mozilla.com/D96360

UltraBlame original commit: 1ebf3b578f0e5f90344424954b171cefac7e8fcd