Bug 1676874 [wpt PR 26504] - [PeerConnection] Improve getStats WPT test coverage., a=testonly
Automatic update from web-platform-tests
[PeerConnection] Improve getStats WPT test coverage.

This CL adds two tests that were discovered missing while reviewing the
test coverage, both of which we PASS:
- getStats(tracks) throws if multiple senders have the same track.
  (Actually it turns out such a test already existed, the old one is
  removed because it was asserting unnecessarily many things.)
- RTCStats.timestamp should increase with time elapsed.

This CL also makes exsting tests more behavior driven. For example, a
test that advertises itself to test that "report containing
peer-connection stats and outbound-track-stats" should not fail because
we have not implemented RTCCodecStats.sdpFmtpLine - this is a different
behavior. As such, tests that are designed to test the existence of a
stats object are updated only to verify that.
  What the old tests were verifying - that the ENTIRE REPORT is valid
and complete - are moved and tested in a separate test that asserts
everything. These tests we still fail because of sdpFmtpLine.

Lastly, the outbound-rtp tests are updated to perform an O/A exchange.
If the senders have not been negotiated to send, there doesn't exist
an encoder or RTP packetizer and as such we shouldn't expect to see
outbound-rtp objects yet.

Bug: chromium:1148286
Change-Id: I50d476f1af41cfae80c77486481e3de06a1d8f5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2534911
Reviewed-by: Philipp Hancke <philipp.hanckegooglemail.com>
Commit-Queue: Henrik Boström <hboschromium.org>
Cr-Commit-Position: refs/heads/master{#827758}

--

wpt-commits: ebcba650d3d41b33f9ca587c7b8416f1f154e6e5
wpt-pr: 26504

UltraBlame original commit: d37d180b0a14c868fd5e05ba328f5992a9706573