Bug 1677825 - Disable hardware acceleration if WebRender fails to initialize on Linux. r=jrmuizel
If we don't disable hardware acceleration when we lose WebRender, we
will likely fallback to OpenGL compositing. This is not a supported
configuration. We already do this step if we decide against using
WebRender during configuration, but not if it was lost at runtime.

Differential Revision: https://phabricator.services.mozilla.com/D97355

UltraBlame original commit: 72e7d30ced00970f4d717f12ce3a01f5d202d1be