Bug 1672967 - [taskgraph.action] Fix bug in actions lookup, r=taskgraph-reviewers,jmaher
According to the documentation, actions are supposed to have an "order" system,
where the first action that applies to a task via its context, is supposed to
take precedence. But when running the match algorithm, we never break out of
the loop. So we end up applying the *last* action that matches a tasks' context.

As far as I can tell, this means we have been using the 'retrigger-disabled'
action for everything and things like retriggering Decision tasks don't work.

Differential Revision: https://phabricator.services.mozilla.com/D97351

UltraBlame original commit: 32f3ce3abd8fa0ca99a0e8fcfaad2167f337fefe